Modularize register map controller
