+++ app.ts
@@ -1,5 +1,6 @@
 import { dtsPare } from "./dts";
 import type { DtsNode, DtsProperty, DtsValue } from "./dts";
+import { createRegisterMap, type RegisterMapController } from "./register-map";
 import {
 	CANVAS_PADDING,
 	NODE_GAP,
@@ -912,384 +913,23 @@ type PropertyLink = PropertyReference;
 const resolvePropertyLinks = (property: DtsProperty): PropertyLink[] =>
 	collectPropertyReferences(property);
 
-type RegisterRange = {
-	node: DtsNode;
-	base: bigint;
-	size: bigint;
-	end: bigint;
-};
-
-type CellSpec = {
-	addressCells: number;
-	sizeCells: number;
-};
-
-const DEFAULT_CELL_SPEC: CellSpec = {
-	addressCells: 1,
-	sizeCells: 1,
-};
-
-const REGISTER_COLOR_PALETTE = [
-	"#2563eb",
-	"#16a34a",
-	"#d97706",
-	"#dc2626",
-	"#7c3aed",
-	"#0891b2",
-	"#f97316",
-	"#0f766e",
-	"#a855f7",
-	"#9333ea",
-];
-
-let registerRanges: RegisterRange[] = [];
-const registerColorByPath = new Map<string, string>();
-let isRegisterMapExpanded = false;
-
-const formatBigIntHex = (value: bigint): string => `0x${value.toString(16)}`;
-
-const formatByteSize = (value: bigint): string => {
-	if (value <= 0n) {
-		return "0 bytes";
-	}
-	const hex = formatBigIntHex(value);
-	if (value > BigInt(Number.MAX_SAFE_INTEGER)) {
-		return hex;
-	}
-	const units = ["bytes", "KiB", "MiB", "GiB", "TiB", "PiB"];
-	let size = Number(value);
-	let unitIndex = 0;
-	while (size >= 1024 && unitIndex < units.length - 1) {
-		size /= 1024;
-		unitIndex += 1;
-	}
-	const formatted = size >= 10 ? size.toFixed(0) : size.toFixed(1);
-	return `${hex} (${formatted} ${units[unitIndex]!})`;
-};
-
-const combineCellsToBigInt = (cells: number[]): bigint => {
-	if (!cells.length) {
-		return 0n;
-	}
-	return cells.reduce((acc, cell) => {
-		const truncated = Math.trunc(cell);
-		const normalized = BigInt.asUintN(32, BigInt(truncated));
-		return (acc << 32n) + normalized;
-	}, 0n);
-};
-
-const normalizeRegValue = (value: DtsValue): number[][] => {
-	const groups: number[][] = [];
-	const pushGroup = (candidate: unknown) => {
-		const numbers = collectNumbersFromValue(candidate as DtsValue);
-		if (numbers.length) {
-			groups.push(numbers);
-		}
-	};
-
-	if (Array.isArray(value)) {
-		if (value.length && Array.isArray(value[0])) {
-			value.forEach((group) => pushGroup(group));
-			return groups;
-		}
-		pushGroup(value);
-		return groups;
-	}
-
-	pushGroup(value);
-	return groups;
-};
-
-const getNumericPropertyValue = (
-	node: DtsNode,
-	name: string,
-): number | null => {
-	const property = node.properties.find((prop) => prop.name === name);
-	if (!property) {
-		return null;
-	}
-	const numbers = collectNumbersFromValue(property.value);
-	if (!numbers.length) {
-		return null;
-	}
-	return numbers[0] ?? null;
-};
-
-const getRegisterColor = (path: string): string => {
-        let color = registerColorByPath.get(path);
-        if (!color) {
-                color =
-                        REGISTER_COLOR_PALETTE[
-                                registerColorByPath.size % REGISTER_COLOR_PALETTE.length
-                        ] ?? "#1d4ed8";
-                registerColorByPath.set(path, color);
-        }
-        return color;
-};
-
-const setRegisterMapExpansion = (expanded: boolean) => {
-        isRegisterMapExpanded = expanded;
-        registerMapPanel?.classList.toggle("collapsed", !expanded);
-        if (registerMapToggle) {
-                registerMapToggle.textContent = expanded
-                        ? "Hide register map"
-                        : "Show register map";
-                registerMapToggle.setAttribute(
-                        "aria-expanded",
-                        expanded ? "true" : "false",
-                );
-        }
-};
-
-const renderRegisterMap = () => {
-	if (
-		!registerMapTrack ||
-		!registerMapLegend ||
-		!registerMapStatus ||
-		!registerMapAxis
-	) {
-		return;
-	}
-
-	registerMapTrack.innerHTML = "";
-	registerMapLegend.innerHTML = "";
-	registerMapAxis.innerHTML = "";
-
-	const hasRanges = registerRanges.length > 0;
-        registerMapTrack.classList.toggle("hidden", !hasRanges);
-        registerMapLegend.classList.toggle("hidden", !hasRanges);
-        registerMapAxis.classList.toggle("hidden", !hasRanges);
-
-        if (registerMapToggle) {
-                registerMapToggle.disabled = !hasRanges;
-        }
-
-        if (!hasRanges) {
-                setRegisterMapExpansion(false);
-                registerMapStatus.textContent = currentRoot
-                        ? "No register ranges detected in this tree."
-                        : "Load a DTS file to view its register map.";
-                return;
-        }
-
-	const minBase = registerRanges.reduce(
-		(min, range) => (range.base < min ? range.base : min),
-		registerRanges[0]!.base,
-	);
-	const maxEnd = registerRanges.reduce(
-		(max, range) => (range.end > max ? range.end : max),
-		registerRanges[0]!.end,
-	);
-	const span = maxEnd > minBase ? maxEnd - minBase : 0n;
-	const totalSpan = span > 0n ? span : 1n;
-	const percentScale = 10000n;
-
-	const regionLabel =
-		registerRanges.length === 1
-			? "1 register region"
-			: `${registerRanges.length} register regions`;
-	const spanText = span > 0n ? formatByteSize(span) : "0 bytes";
-	registerMapStatus.textContent = `${regionLabel}. Address span ${formatBigIntHex(
-		minBase,
-	)} – ${formatBigIntHex(maxEnd)} (${spanText}).`;
-
-	const axisStart = document.createElement("span");
-	axisStart.textContent = formatBigIntHex(minBase);
-	const axisEnd = document.createElement("span");
-	axisEnd.textContent = formatBigIntHex(maxEnd);
-	registerMapAxis.append(axisStart, axisEnd);
-
-	let cursor = minBase;
-	registerRanges.forEach((range) => {
-		if (range.base > cursor) {
-			const gap = range.base - cursor;
-			const gapUnits = Number((gap * percentScale) / totalSpan);
-			if (gapUnits > 0) {
-				const gapPercent = gapUnits / 100;
-				const gapElement = document.createElement("div");
-				gapElement.className = "register-gap";
-				gapElement.style.flex = `0 0 ${gapPercent}%`;
-				registerMapTrack.append(gapElement);
-			}
-		}
-
-		const layoutSize = range.size > 0n ? range.size : 0n;
-		const widthUnits =
-			layoutSize > 0n
-				? Number((layoutSize * percentScale) / totalSpan)
-				: 0;
-		const widthPercent = layoutSize > 0n ? widthUnits / 100 : 0;
-		const segment = document.createElement("button");
-		segment.type = "button";
-		segment.className = "register-segment";
-		if (selectedNodePath === range.node.path) {
-			segment.classList.add("selected");
-		}
-		const color = getRegisterColor(range.node.path);
-		segment.style.backgroundColor = color;
-		segment.style.minWidth = "6px";
-		if (layoutSize > 0n) {
-			segment.style.flex = `0 0 ${widthPercent}%`;
-		} else {
-			segment.style.flex = "0 0 auto";
-			segment.style.width = "6px";
-		}
-		const label = range.node.label
-			? `${range.node.label}: ${range.node.fullName}`
-			: range.node.fullName;
-		const end = range.size > 0n ? range.end : range.base;
-		const sizeText = formatByteSize(range.size);
-		segment.title =
-			range.size > 0n
-				? `${label} — ${formatBigIntHex(range.base)} to ${formatBigIntHex(
-						end,
-					)} (${sizeText})`
-				: `${label} — ${formatBigIntHex(range.base)} (size: 0)`;
-		segment.addEventListener("click", () => {
-			focusNodeByPath(range.node.path);
-		});
-		registerMapTrack.append(segment);
-
-		if (range.end > cursor) {
-			cursor = range.end;
-		}
-	});
-
-	const legendTable = document.createElement("table");
-	const thead = document.createElement("thead");
-	const headerRow = document.createElement("tr");
-	["", "Node", "Start", "End", "Size"].forEach((heading) => {
-		const th = document.createElement("th");
-		th.textContent = heading;
-		headerRow.append(th);
-	});
-	thead.append(headerRow);
-	legendTable.append(thead);
-
-	const tbody = document.createElement("tbody");
-	registerRanges.forEach((range) => {
-		const row = document.createElement("tr");
-		if (selectedNodePath === range.node.path) {
-			row.classList.add("selected");
-		}
-
-		const swatchCell = document.createElement("td");
-		const swatch = document.createElement("span");
-		swatch.className = "register-swatch";
-		swatch.style.backgroundColor = getRegisterColor(range.node.path);
-		swatchCell.append(swatch);
-
-		const nodeCell = document.createElement("td");
-		const nodeButton = document.createElement("button");
-		nodeButton.type = "button";
-		nodeButton.className = "register-link";
-		nodeButton.textContent = range.node.label
-			? `${range.node.label}: ${range.node.fullName}`
-			: range.node.fullName;
-		nodeButton.addEventListener("click", () => {
-			focusNodeByPath(range.node.path);
-		});
-		nodeCell.append(nodeButton);
-
-		const startCell = document.createElement("td");
-		startCell.textContent = formatBigIntHex(range.base);
-
-		const endCell = document.createElement("td");
-		endCell.textContent = formatBigIntHex(
-			range.size > 0n ? range.end : range.base,
-		);
-
-		const sizeCell = document.createElement("td");
-		sizeCell.textContent = formatByteSize(range.size);
-
-		row.append(swatchCell, nodeCell, startCell, endCell, sizeCell);
-		tbody.append(row);
-	});
-
-	legendTable.append(tbody);
-	registerMapLegend.append(legendTable);
-};
-
-const computeRegisterRanges = (root: DtsNode | null) => {
-	registerRanges = [];
-	registerColorByPath.clear();
-	if (!root) {
-		renderRegisterMap();
-		return;
-	}
-
-	const ranges: RegisterRange[] = [];
-	const visit = (node: DtsNode, parentSpec: CellSpec) => {
-		const regProperty = node.properties.find((prop) => prop.name === "reg");
-		if (regProperty) {
-			const groups = normalizeRegValue(regProperty.value);
-			const addressCells = Math.max(0, parentSpec.addressCells);
-			const sizeCells = Math.max(0, parentSpec.sizeCells);
-			const chunkSize = addressCells + sizeCells;
-			if (chunkSize > 0) {
-				groups.forEach((group) => {
-					if (group.length < chunkSize) {
-						return;
-					}
-					for (
-						let offset = 0;
-						offset + chunkSize <= group.length;
-						offset += chunkSize
-					) {
-						const addressSlice = group.slice(
-							offset,
-							offset + addressCells,
-						);
-						const sizeSlice = group.slice(
-							offset + addressCells,
-							offset + chunkSize,
-						);
-						const base = combineCellsToBigInt(addressSlice);
-						const size =
-							sizeCells > 0
-								? combineCellsToBigInt(sizeSlice)
-								: 0n;
-						const end = size > 0n ? base + size : base;
-						ranges.push({ node, base, size, end });
-					}
-				});
-			}
-		}
-
-		const nextSpec: CellSpec = {
-			addressCells:
-				getNumericPropertyValue(node, "#address-cells") ??
-				parentSpec.addressCells,
-			sizeCells:
-				getNumericPropertyValue(node, "#size-cells") ??
-				parentSpec.sizeCells,
-		};
-
-		node.children.forEach((child) => visit(child, nextSpec));
-	};
-
-	visit(root, DEFAULT_CELL_SPEC);
-
-	ranges.sort((a, b) => {
-		if (a.base === b.base) {
-			if (a.size === b.size) {
-				return a.node.path.localeCompare(b.node.path);
-			}
-			return a.size < b.size ? -1 : 1;
-		}
-		return a.base < b.base ? -1 : 1;
-	});
-
-	registerRanges = ranges;
-	renderRegisterMap();
-};
+const registerMap: RegisterMapController = createRegisterMap({
+        panel: registerMapPanel,
+        toggle: registerMapToggle,
+        status: registerMapStatus,
+        track: registerMapTrack,
+        legend: registerMapLegend,
+        axis: registerMapAxis,
+        onFocusNode: focusNodeByPath,
+        collectNumbersFromValue,
+});
 
 const renderDetails = (node: DtsNode | null) => {
 	if (!detailsContent || !detailsTitle) {
 		return;
 	}
 
-	renderRegisterMap();
+        registerMap.setSelection(node?.path ?? null);
 
 	detailsContent.innerHTML = "";
 
@@ -1717,7 +1357,7 @@ const displayTree = (source: string, origin: string) => {
 		selectedNodePath = null;
 		selectedNode = null;
 		rebuildNodeIndexes(null);
-		computeRegisterRanges(null);
+                registerMap.updateRanges(null);
 		renderDetails(null);
 		return;
 	}
@@ -1729,7 +1369,7 @@ const displayTree = (source: string, origin: string) => {
 	activeFilterRaw = "";
 	activeFilterNormalized = "";
 	rebuildNodeIndexes(currentRoot);
-	computeRegisterRanges(currentRoot);
+        registerMap.updateRanges(currentRoot);
 	renderDetails(null);
 	if (canvas) {
 		canvas.style.cursor = "default";
@@ -2141,10 +1781,6 @@ const attachEventHandlers = () => {
                 displayTree(SAMPLE_DTS, "Sample DTS");
         });
 
-        registerMapToggle?.addEventListener("click", () => {
-                setRegisterMapExpansion(!isRegisterMapExpanded);
-        });
-
         warningsToggle?.addEventListener("change", () => {
                 renderWarningsPanel();
                 if (lastStatus) {
@@ -2433,14 +2069,13 @@ const attachEventHandlers = () => {
 	});
 };
 
-setRegisterMapExpansion(false);
+registerMap.setExpanded(false);
 
 if (!ctx) {
-	statusElement?.classList.add("error");
-	statusElement?.append(
-		"\nCanvas rendering context unavailable in this browser.",
-	);
+        statusElement?.classList.add("error");
+        statusElement?.append(
+                "\nCanvas rendering context unavailable in this browser.",
+        );
 } else {
-	attachEventHandlers();
-	renderRegisterMap();
+        attachEventHandlers();
 }
